+++ hot-refresh-core/src/test/resources/com/hyf/hotrefresh/generate/ExtraClass.java
@@ -1,0 +1,17 @@
+package com.hyf.hotrefresh.generate;
+
+import java.util.List;
+
+/**
+ * @author baB_hyf
+ * @date 2022/06/25
+ */
+public class ExtraClass {
+
+    private Integer      i;
+    private List<String> list;
+
+    public String getString() {
+        return "extraString";
+    }
+}